Don't copy request parameters to tab links automatically.
In the first public release of this module query parameters from the
current request were appended to the tab URIs. The idea behind it was to
easily pass a session id. This turned out to be a bad idea, because all
kinds of paramaters were passed to pages, where they might introduce
complications.  Therefore beginning with this release no query parameters
are appended to the tab URIs automatically, and the desired query
parameters must be passed manually in the "BUILD_TABS" callback method.